update result names and add to plot methods
